Move sanitize so we don't import the world into every bundle
