+++ src/js/common/schemaform/helpers.js
@@ -542,17 +542,3 @@ export function getActiveChapters(formConfig, formData) {
   return _.uniq(expandedPageList.map(p => p.chapterKey).filter(key => !!key && key !== 'review'));
 }
 
-export function sanitizeForm(formData) {
-  try {
-    const suffixes = ['vaFileNumber', 'first', 'last', 'accountNumber', 'socialSecurityNumber', 'dateOfBirth'];
-    return JSON.stringify(formData, (key, value) => {
-      if (value && suffixes.some(suffix => key.toLowerCase().endsWith(suffix.toLowerCase()))) {
-        return 'removed';
-      }
-
-      return value;
-    });
-  } catch (e) {
-    return null;
-  }
-}
